+++ routes/incomes.js
@@ -8,10 +8,13 @@ const {incomeDB} = require('./../dataBase/db');
 router.post('/add' ,authenticateToken,(req,res) =>{
 
   const { description, amount, category} = req.body;
+  if (!description || !amount || !category) {
+    return res.status(400).json({ error: 'Missing required fields' });
+  }
   const userId = req.user.id;
   const date = new Date().toISOString().slice(0, 10);
   incomeDB.run(
-      `(description, amount, category, date, user_id) VALUES (?, ?, ?, ?, ?)`,
+      `INSERT INTO incomes (description, amount, category, date, user_id) VALUES (?, ?, ?, ?, ?)`,
       [description, amount, category, date, userId],
       function (err) {
         if (err) {
@@ -36,6 +39,9 @@ router.get('/',authenticateToken,(req, res) => {
   router.delete('/delete/:id', authenticateToken, (req, res) => {
     const userId = req.user.id;
     const { id } = req.params;
+    if (!id) {
+      return res.status(400).json({ error: 'Missing ID' });
+    }
     expensesDB.run(`DELETE FROM incomes WHERE user_id = ? AND id = ?`, [userId, id],
       function (err) {  
         if (err) {
@@ -53,7 +59,13 @@ router.get('/',authenticateToken,(req, res) => {
   
     const userId = req.user.id;
     const { id } = req.params;
+    if (!id) {
+      return res.status(400).json({ error: 'Missing ID' });
+    }
     const { description, amount, category } = req.body;
+    if (!description || !amount || !category) {
+      return res.status(400).json({ error: 'Missing required fields' });
+    }
     expensesDB.run(
       `UPDATE incomes SET description = ?, amount = ?, category = ? WHERE user_id = ? AND  id = ?`,
       [description, amount, category, userId, id],
